requirements.txt
# Core Agent Framework dependencies
agent-framework-azure-ai
azure-identity

# Observability dependencies
agent-framework[telemetry]
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-grpc

# Optional: For local development and testing
opentelemetry-instrumentation